refactor to viewmodel file
